cli/requirements.txt
fastapi
typer
starlette
uvicorn
httpx
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
python-multipart